Refactor: Change layout appropriate for the member domain #23
